fixed bug with L2 distance

+++ src/main/scala/cs231n/data/LabeledImage.scala
@@ -33,10 +33,10 @@ class LabeledImage(labelValue: Int, redValues: Array[Int], greenValues: Array[In
   }
 
   // L2 (Euclidean) Distance function
-  def l2Distance(image: LabeledImage): Int = { // Zip image distances together with absolute difference
+  def l2Distance(image: LabeledImage): Double = { // Zip image distances together with absolute difference
     Math.sqrt((this.data zip image.data).map(x => x match { // merge the zipped values together
       case (a, b) => Math.sqrt(b - a) // Absolute value of the distance
-    }).fold(0)(_ + _)) // sum the deltas
+    }).fold(0.0)(_ + _)) // sum the deltas
   }
 
   // Label accessor
